Renaming cso starter package
