+++ src/backend/libs/efiling-cso-starter/src/test/java/ca/bc/gov/open/jag/efilingcsostarter/config/AutoConfigurationTest.java
@@ -1,11 +1,12 @@
-package ca.bc.gov.open.jag.efilingaccountclient.config;
+package ca.bc.gov.open.jag.efilingcsostarter.config;
 
 
-import ca.bc.gov.open.jag.efilingaccountclient.CsoAccountServiceImpl;
+import ca.bc.gov.open.jag.efilingcsostarter.CsoAccountServiceImpl;
 import ca.bc.gov.open.jag.efilingaccountclient.mappers.AccountDetailsMapperImpl;
 import ca.bc.gov.open.jag.efilingcommons.model.Clients;
 import ca.bc.gov.open.jag.efilingcommons.model.EfilingSoapClientProperties;
 import ca.bc.gov.open.jag.efilingcommons.model.SoapProperties;
+import ca.bc.gov.open.jag.efilingcsostarter.config.AutoConfiguration;
 import org.junit.jupiter.api.*;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
